+++ app/Main.hs
@@ -2,11 +2,38 @@ module Main where
 
 import ClassyPrelude
 import Dobby.ChangeLog
+import Options.Applicative
+import Options.Applicative.Text
+
+commitOpts :: Parser Changes
+commitOpts = Changes
+  <$> many (textOption
+    ( long "added"
+   <> short 'a'
+   <> metavar "ADDED" ) )
+  <*> many (textOption
+    ( long "changed"
+   <> short 'c'
+   <> metavar "CHANGED" ) )
+  <*> many (textOption
+    ( long "deprecated"
+   <> short 'd'
+   <> metavar "DEPRECATED" ) )
+  <*> many (textOption
+    ( long "fixed"
+   <> short 'f'
+   <> metavar "FIXED" ) )
+  <*> many (textOption
+    ( long "Removed"
+   <> short 'r'
+   <> metavar "REMOVED" ) )
+  <*> many (textOption
+    ( long "Security"
+   <> short 's'
+   <> metavar "SECURITY" ) )
 
 main :: IO ()
-main = do
-  -- TODO: print usage
-  Just tool <- listToMaybe . toList <$> getArgs
-  case tool of
-    "bump-patch-version" -> patchVersion
-    _ -> putStrLn $ "Unrecognized tool: " ++ tool
+main =
+  join . execParser . flip info idm $ subparser
+    ( command "commit" (info (commit <$> commitOpts) idm)
+   <> command "patch-release" (info (pure patchVersion) idm) )